agents/todo_list/pyproject.toml
[project]
name = "todo-list-agent"
version = "0.1.0"
description = "A slightly personalised todo list agent for Shinox mesh"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0.0",
    "faststream[cli,kafka]>=0.6.5",
    "mcp>=1.26.0",
    "shinox-agent-sdk>=0.1.1",
]